+++ banyand/liaison/grpc/discovery.go
@@ -18,7 +18,10 @@
 package grpc
 
 import (
+	"context"
+	"fmt"
 	"sync"
+	"time"
 
 	"github.com/pkg/errors"
 
@@ -26,9 +29,10 @@ import (
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
+	"github.com/apache/skywalking-banyandb/banyand/metadata"
+	"github.com/apache/skywalking-banyandb/banyand/metadata/schema"
 	"github.com/apache/skywalking-banyandb/banyand/queue"
 	"github.com/apache/skywalking-banyandb/banyand/tsdb"
-	"github.com/apache/skywalking-banyandb/pkg/bus"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
 	"github.com/apache/skywalking-banyandb/pkg/partition"
 )
@@ -36,18 +40,99 @@ import (
 var errNotExist = errors.New("the object doesn't exist")
 
 type discoveryService struct {
-	shardRepo  *shardRepo
-	entityRepo *entityRepo
-	pipeline   queue.Queue
-	log        *logger.Logger
+	pipeline     queue.Queue
+	metadataRepo metadata.Repo
+	shardRepo    *shardRepo
+	entityRepo   *entityRepo
+	log          *logger.Logger
+	kind         schema.Kind
 }
 
-func newDiscoveryService(pipeline queue.Queue) *discoveryService {
+func newDiscoveryService(pipeline queue.Queue, kind schema.Kind, metadataRepo metadata.Repo) *discoveryService {
+	sr := &shardRepo{shardEventsMap: make(map[identity]uint32)}
+	er := &entityRepo{entitiesMap: make(map[identity]partition.EntityLocator)}
 	return &discoveryService{
-		shardRepo:  &shardRepo{shardEventsMap: make(map[identity]uint32)},
-		entityRepo: &entityRepo{entitiesMap: make(map[identity]partition.EntityLocator)},
-		pipeline:   pipeline,
+		shardRepo:    sr,
+		entityRepo:   er,
+		pipeline:     pipeline,
+		kind:         kind,
+		metadataRepo: metadataRepo,
 	}
+}
+
+func (ds *discoveryService) initialize(ctx context.Context) error {
+	ctxLocal, cancel := context.WithTimeout(ctx, 5*time.Second)
+	groups, err := ds.metadataRepo.GroupRegistry().ListGroup(ctxLocal)
+	cancel()
+	if err != nil {
+		return err
+	}
+	for _, g := range groups {
+		switch ds.kind {
+		case schema.KindMeasure:
+		case schema.KindStream:
+		default:
+			continue
+		}
+		ctxLocal, cancel := context.WithTimeout(ctx, 5*time.Second)
+		shards, innerErr := ds.metadataRepo.ShardRegistry().ListShard(ctxLocal, schema.ListOpt{Group: g.Metadata.Name})
+		cancel()
+		if innerErr != nil {
+			return innerErr
+		}
+		for _, s := range shards {
+			ds.shardRepo.OnAddOrUpdate(schema.Metadata{
+				TypeMeta: schema.TypeMeta{
+					Kind:  schema.KindShard,
+					Name:  s.Metadata.Name,
+					Group: s.Metadata.Group,
+				},
+				Spec: s,
+			})
+		}
+
+		switch ds.kind {
+		case schema.KindMeasure:
+			ctxLocal, cancel = context.WithTimeout(ctx, 5*time.Second)
+			mm, innerErr := ds.metadataRepo.MeasureRegistry().ListMeasure(ctxLocal, schema.ListOpt{Group: g.Metadata.Name})
+			cancel()
+			if innerErr != nil {
+				return innerErr
+			}
+			for _, m := range mm {
+				ds.entityRepo.OnAddOrUpdate(schema.Metadata{
+					TypeMeta: schema.TypeMeta{
+						Kind:  schema.KindMeasure,
+						Name:  m.Metadata.Name,
+						Group: m.Metadata.Group,
+					},
+					Spec: m,
+				})
+			}
+		case schema.KindStream:
+			ctxLocal, cancel = context.WithTimeout(ctx, 5*time.Second)
+			ss, innerErr := ds.metadataRepo.StreamRegistry().ListStream(ctxLocal, schema.ListOpt{Group: g.Metadata.Name})
+			cancel()
+			if innerErr != nil {
+				return innerErr
+			}
+			for _, s := range ss {
+				ds.entityRepo.OnAddOrUpdate(schema.Metadata{
+					TypeMeta: schema.TypeMeta{
+						Kind:  schema.KindStream,
+						Name:  s.Metadata.Name,
+						Group: s.Metadata.Group,
+					},
+					Spec: s,
+				})
+			}
+		default:
+			return fmt.Errorf("unsupported kind: %d", ds.kind)
+		}
+	}
+	ds.metadataRepo.RegisterHandler(schema.KindShard, ds.shardRepo)
+	ds.metadataRepo.RegisterHandler(ds.kind, ds.entityRepo)
+	return nil
 }
 
 func (ds *discoveryService) SetLogger(log *logger.Logger) {
@@ -75,38 +160,46 @@ type identity struct {
 	group string
 }
 
+func (i identity) String() string {
+	return fmt.Sprintf("%s/%s", i.group, i.name)
+}
+
+var _ schema.EventHandler = (*shardRepo)(nil)
+
 type shardRepo struct {
 	log            *logger.Logger
 	shardEventsMap map[identity]uint32
 	sync.RWMutex
 }
 
-func (s *shardRepo) Rev(message bus.Message) (resp bus.Message) {
-	e, ok := message.Data().(*databasev1.ShardEvent)
-	if !ok {
-		s.log.Warn().Msg("invalid e data type")
+// OnAddOrUpdate implements schema.EventHandler.
+func (s *shardRepo) OnAddOrUpdate(schemaMetadata schema.Metadata) {
+	if schemaMetadata.Kind != schema.KindShard {
 		return
 	}
-	s.setShardNum(e)
-
+	shard := schemaMetadata.Spec.(*databasev1.Shard)
+	idx := getID(shard.GetMetadata())
 	if le := s.log.Debug(); le.Enabled() {
-		le.
-			Str("action", databasev1.Action_name[int32(e.Action)]).
-			Uint64("shardID", e.Shard.Id).
-			Msg("received a shard e")
+		le.Stringer("id", idx).Uint32("total", shard.Total).Msg("shard added or updated")
 	}
-	return
-}
-
-func (s *shardRepo) setShardNum(eventVal *databasev1.ShardEvent) {
 	s.RWMutex.Lock()
 	defer s.RWMutex.Unlock()
-	idx := getID(eventVal.GetShard().GetMetadata())
-	if eventVal.Action == databasev1.Action_ACTION_PUT {
-		s.shardEventsMap[idx] = eventVal.Shard.Total
-	} else if eventVal.Action == databasev1.Action_ACTION_DELETE {
-		delete(s.shardEventsMap, idx)
+	s.shardEventsMap[idx] = shard.Total
+}
+
+// OnDelete implements schema.EventHandler.
+func (s *shardRepo) OnDelete(schemaMetadata schema.Metadata) {
+	if schemaMetadata.Kind != schema.KindShard {
+		return
 	}
+	shard := schemaMetadata.Spec.(*databasev1.Shard)
+	idx := getID(shard.GetMetadata())
+	if le := s.log.Debug(); le.Enabled() {
+		le.Stringer("id", idx).Msg("shard deleted")
+	}
+	s.RWMutex.Lock()
+	defer s.RWMutex.Unlock()
+	delete(s.shardEventsMap, idx)
 }
 
 func (s *shardRepo) shardNum(idx identity) (uint32, bool) {
@@ -126,49 +219,96 @@ func getID(metadata *commonv1.Metadata) identity {
 	}
 }
 
+var _ schema.EventHandler = (*entityRepo)(nil)
+
 type entityRepo struct {
 	log         *logger.Logger
 	entitiesMap map[identity]partition.EntityLocator
 	sync.RWMutex
 }
 
-func (s *entityRepo) Rev(message bus.Message) (resp bus.Message) {
-	e, ok := message.Data().(*databasev1.EntityEvent)
-	if !ok {
-		s.log.Warn().Msg("invalid e data type")
+// OnAddOrUpdate implements schema.EventHandler.
+func (e *entityRepo) OnAddOrUpdate(schemaMetadata schema.Metadata) {
+	var el partition.EntityLocator
+	var id identity
+	switch schemaMetadata.Kind {
+	case schema.KindMeasure:
+		measure := schemaMetadata.Spec.(*databasev1.Measure)
+		el = partition.NewEntityLocator(measure.TagFamilies, measure.Entity)
+		id = getID(measure.GetMetadata())
+	case schema.KindStream:
+		stream := schemaMetadata.Spec.(*databasev1.Stream)
+		el = partition.NewEntityLocator(stream.TagFamilies, stream.Entity)
+		id = getID(stream.GetMetadata())
+	default:
 		return
 	}
-	id := getID(e.GetSubject())
-	if le := s.log.Debug(); le.Enabled() {
-		le.
-			Str("action", databasev1.Action_name[int32(e.Action)]).
-			Interface("subject", id).
-			Msg("received an entity event")
-	}
-	s.RWMutex.Lock()
-	defer s.RWMutex.Unlock()
-	switch e.Action {
-	case databasev1.Action_ACTION_PUT:
-		en := make(partition.EntityLocator, 0, len(e.GetEntityLocator()))
-		for _, l := range e.GetEntityLocator() {
-			en = append(en, partition.TagLocator{
-				FamilyOffset: int(l.FamilyOffset),
-				TagOffset:    int(l.TagOffset),
-			})
+	if le := e.log.Debug(); le.Enabled() {
+		var kind string
+		switch schemaMetadata.Kind {
+		case schema.KindMeasure:
+			kind = "measure"
+		case schema.KindStream:
+			kind = "stream"
+		default:
+			kind = "unknown"
 		}
-		s.entitiesMap[id] = en
-	case databasev1.Action_ACTION_DELETE:
-		delete(s.entitiesMap, id)
-	case databasev1.Action_ACTION_UNSPECIFIED:
-		s.log.Warn().RawJSON("event", logger.Proto(e)).Msg("ignored unspecified event")
+		le.
+			Str("action", "add_or_update").
+			Stringer("subject", id).
+			Str("kind", kind).
+			Msg("entity added or updated")
 	}
-	return
+	en := make(partition.EntityLocator, 0, len(el))
+	for _, l := range el {
+		en = append(en, partition.TagLocator{
+			FamilyOffset: l.FamilyOffset,
+			TagOffset:    l.TagOffset,
+		})
+	}
+	e.RWMutex.Lock()
+	defer e.RWMutex.Unlock()
+	e.entitiesMap[id] = en
 }
 
-func (s *entityRepo) getLocator(id identity) (partition.EntityLocator, bool) {
-	s.RWMutex.RLock()
-	defer s.RWMutex.RUnlock()
-	el, ok := s.entitiesMap[id]
+// OnDelete implements schema.EventHandler.
+func (e *entityRepo) OnDelete(schemaMetadata schema.Metadata) {
+	var id identity
+	switch schemaMetadata.Kind {
+	case schema.KindMeasure:
+		measure := schemaMetadata.Spec.(*databasev1.Measure)
+		id = getID(measure.GetMetadata())
+	case schema.KindStream:
+		stream := schemaMetadata.Spec.(*databasev1.Stream)
+		id = getID(stream.GetMetadata())
+	default:
+		return
+	}
+	if le := e.log.Debug(); le.Enabled() {
+		var kind string
+		switch schemaMetadata.Kind {
+		case schema.KindMeasure:
+			kind = "measure"
+		case schema.KindStream:
+			kind = "stream"
+		default:
+			kind = "unknown"
+		}
+		le.
+			Str("action", "delete").
+			Stringer("subject", id).
+			Str("kind", kind).
+			Msg("entity deleted")
+	}
+	e.RWMutex.Lock()
+	defer e.RWMutex.Unlock()
+	delete(e.entitiesMap, id)
+}
+
+func (e *entityRepo) getLocator(id identity) (partition.EntityLocator, bool) {
+	e.RWMutex.RLock()
+	defer e.RWMutex.RUnlock()
+	el, ok := e.entitiesMap[id]
 	if !ok {
 		return nil, false
 	}