+++ banyand/stream/service.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/apache/skywalking-banyandb/api/data"
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
-	"github.com/apache/skywalking-banyandb/banyand/discovery"
+	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
 	"github.com/apache/skywalking-banyandb/banyand/metadata"
 	"github.com/apache/skywalking-banyandb/banyand/metadata/schema"
 	"github.com/apache/skywalking-banyandb/banyand/observability"
@@ -55,10 +55,8 @@ type service struct {
 	schemaRepo      schemaRepo
 	metadata        metadata.Repo
 	pipeline        queue.Queue
-	repo            discovery.ServiceRepo
 	writeListener   *writeCallback
 	l               *logger.Logger
-	stopCh          chan struct{}
 	root            string
 	dbOpts          tsdb.DatabaseOpts
 	blockBufferSize run.Bytes
@@ -96,17 +94,21 @@ func (s *service) Name() string {
 	return "stream"
 }
 
-func (s *service) PreRun() error {
+func (s *service) Role() databasev1.Role {
+	return databasev1.Role_ROLE_DATA
+}
+
+func (s *service) PreRun(ctx context.Context) error {
 	s.l = logger.GetLogger(s.Name())
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	groups, err := s.metadata.GroupRegistry().ListGroup(ctx)
+	ctxLocal, cancel := context.WithTimeout(ctx, 5*time.Second)
+	groups, err := s.metadata.GroupRegistry().ListGroup(ctxLocal)
 	cancel()
 	if err != nil {
 		return err
 	}
 	path := path.Join(s.root, s.Name())
 	observability.UpdatePath(path)
-	s.schemaRepo = newSchemaRepo(path, s.metadata, s.repo, int64(s.blockBufferSize), s.dbOpts, s.l)
+	s.schemaRepo = newSchemaRepo(path, s.metadata, int64(s.blockBufferSize), s.dbOpts, s.l)
 	for _, g := range groups {
 		if g.Catalog != commonv1.Catalog_CATALOG_STREAM {
 			continue
@@ -115,18 +117,22 @@ func (s *service) PreRun() error {
 		if err != nil {
 			return err
 		}
-		ctx, cancel = context.WithTimeout(context.Background(), 5*time.Second)
-		schemas, err := s.metadata.StreamRegistry().ListStream(ctx, schema.ListOpt{Group: gp.GetSchema().GetMetadata().Name})
+		ctxLocal, cancel := context.WithTimeout(ctx, 5*time.Second)
+		schemas, err := s.metadata.StreamRegistry().ListStream(ctxLocal, schema.ListOpt{Group: gp.GetSchema().GetMetadata().Name})
 		cancel()
 		if err != nil {
 			return err
 		}
 		for _, sa := range schemas {
-			if _, innerErr := gp.StoreResource(sa); innerErr != nil {
+			if _, innerErr := gp.StoreResource(ctx, sa); innerErr != nil {
 				return innerErr
 			}
 		}
 	}
+	// run a serial watcher
+	s.schemaRepo.Watcher()
+	s.metadata.RegisterHandler(schema.KindGroup|schema.KindStream|schema.KindIndexRuleBinding|schema.KindIndexRule,
+		&s.schemaRepo)
 
 	s.writeListener = setUpWriteCallback(s.l, &s.schemaRepo)
 
@@ -138,32 +144,21 @@ func (s *service) PreRun() error {
 }
 
 func (s *service) Serve() run.StopNotify {
-	_ = s.schemaRepo.NotifyAll()
-	// run a serial watcher
-	s.schemaRepo.Watcher()
-
-	s.metadata.StreamRegistry().RegisterHandler(schema.KindGroup|schema.KindStream|schema.KindIndexRuleBinding|schema.KindIndexRule,
-		&s.schemaRepo)
-	return s.stopCh
+	return s.schemaRepo.StopCh()
 }
 
 func (s *service) GracefulStop() {
 	s.schemaRepo.Close()
-	if s.stopCh != nil {
-		close(s.stopCh)
-	}
 }
 
 // NewService returns a new service.
-func NewService(_ context.Context, metadata metadata.Repo, repo discovery.ServiceRepo, pipeline queue.Queue) (Service, error) {
+func NewService(_ context.Context, metadata metadata.Repo, pipeline queue.Queue) (Service, error) {
 	return &service{
 		metadata: metadata,
-		repo:     repo,
 		pipeline: pipeline,
 		dbOpts: tsdb.DatabaseOpts{
 			EnableGlobalIndex: true,
 			IndexGranularity:  tsdb.IndexGranularityBlock,
 		},
-		stopCh: make(chan struct{}),
 	}, nil
 }