+++ CHANGES.md
@@ -10,6 +10,8 @@ Release Notes.
 - Implement Write-ahead Logging
 - Document the clustering.
 - Support multiple roles for banyand server.
+- Support for recovery buffer using wal.
+- Register the node role to the metadata registry.
 
 ### Bugs
 
@@ -22,6 +24,7 @@ Release Notes.
 ### Chores
 
 - Bump several dependencies and tools.
+- Drop redundant "discovery" module from banyand. "metadata" module is enough to play the node and shard discovery role.
 
 ## 0.4.0